♻ refactor: 상세 페이지 setting 분리
